# docs/swagger.yaml
definitions:
...
  dbmodels.Lookup:
    properties:
      actionType:
        type: integer
      approvalStatus:
        type: integer
      code:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      descr:
        type: string
      id:
        type: integer
      isAlternateEntry:
        type: boolean
      isHighRisk:
        type: boolean
      latestApproval:
        type: string
      latestApprover:
        type: string
      latestSuggestion:
        type: string
      latestSuggestor:
        type: string
      lookupGroup:
        type: string
      name:
        type: string
      orderNo:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
      version:
        type: integer
    type: object
...
  dbmodels.MasterTag:
    properties:
      code:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  dbmodels.PartnerLink:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      merchant_id:
        type: string
      partner_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ReqLookupDto:
    properties:
      code:
        type: string
      descr:
        type: string
      id:
        type: integer
      limit:
        type: integer
      lookupGroup:
        type: string
      name:
        type: string
      orderNo:
        type: integer
      page:
        type: integer
    type: object
...
  dto.ReqMasterTagDto:
    properties:
      code:
        example: "00001"
        type: string
      description:
        example: nama nama ikan
        type: string
      id:
        type: integer
      limit:
        example: 10
        type: integer
      name:
        example: sampo
        type: string
      page:
        example: 1
        type: integer
      status:
        example: active
        type: string
    type: object
...
  dto.ReqPartnerLinkDto:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      limit:
        type: integer
      merchant_id:
        type: string
      page:
        type: integer
      partner_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Response:
    properties:
      contents: {}
      data: {}
      errCode:
        type: string
      errDesc:
        type: string
      totalData:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /rosego/v.0.1/lookup:
    post:
      consumes:
      - application/json
      description: Lookup Save/EDit
      operationId: LookupSave/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqLookupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Lookup - Save / Edit
...
  /rosego/v.0.1/lookup/filter:
    post:
      consumes:
      - application/json
      description: Lookup Filter Paging
      operationId: LookupFilter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqLookupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.Lookup'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Lookup - Filter
  /rosego/v.0.1/master-tag:
    post:
      consumes:
      - application/json
      description: MasterTag Save/EDit
      operationId: MasterTag - Save/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqMasterTagDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: MasterTag - Save / Edit
  /rosego/v.0.1/master-tag/all:
    get:
      consumes:
      - application/json
      description: MasterTag AllData
      operationId: MasterTagAllData
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.MasterTag'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: MasterTag - All
  /rosego/v.0.1/master-tag/filter:
    post:
      consumes:
      - application/json
      description: MasterTag Filter Paging
      operationId: MasterTagFilter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqMasterTagDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.MasterTag'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: MasterTag - Filter
...
  /rosego/v.0.1/partner-Link:
    post:
      consumes:
      - application/json
      description: PartnerLinkController Save/EDit
      operationId: PartnerLinkController - Save/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqPartnerLinkDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: PartnerLinkController - Save / Edit
  /rosego/v.0.1/partner-link/delete/:
    post:
      consumes:
      - application/json
      description: PartnerLinkController Delete
      operationId: PartnerLinkController - Delete
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqPartnerLinkDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: PartnerLinkController - Delete
  /rosego/v.0.1/partner-link/filter:
    post:
      consumes:
      - application/json
      description: PartnerLink Filter Paging
      operationId: PartnerLink - Filter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqPartnerLinkDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.PartnerLink'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: PartnerLink - Filter
...
swagger: "2.0"